feat: update EC2 Resource Detector to IDMSv2 and modified detect-resource
