+++ packages/opentelemetry-resources/src/platform/node/detectors/index.ts
@@ -17,4 +17,3 @@
 export { awsEc2Detector } from './AwsEc2Detector';
 export { envDetector } from './EnvDetector';
 export { gcpDetector } from './GcpDetector';
-export { awsEc2DetectorTemp } from './AwsEc2DetectorTemp';
